Code sharing between crawl and runspider command
